+++ frontend/package.json
@@ -12,6 +12,7 @@
     "@types/react-dom": "^17.0.0",
     "bootstrap": "^5.0.2",
     "react": "^17.0.2",
+    "react-bootstrap": "^1.6.1",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
