+++ frontend/src/pages/Dashboard.tsx
@@ -1,0 +1,9 @@
+function Dashboard() {
+  return (
+    <div>
+      <h1>Merchant dashboard</h1>
+    </div>
+  );
+}
+
+export default Dashboard;
