+++ frontend/src/pages/Landing.tsx
@@ -1,0 +1,14 @@
+import logo from "../logo.svg";
+
+function Landing() {
+  return (
+    <div>
+      <header className="App-header">
+        <embed src={logo} className="App-logo" />
+        <h2>Welcome to Tiden</h2>
+      </header>
+    </div>
+  );
+}
+
+export default Landing;
