+++ frontend/src/pages/MerchantProfile.tsx
@@ -1,0 +1,11 @@
+import PaymentForm from "../components/PaymentForm";
+
+function MerchantProfile() {
+  return (
+    <div>
+      <PaymentForm />
+    </div>
+  );
+}
+
+export default MerchantProfile;
